feat: added bottom sheet
- added bottom sheet package
- added dependencies for bottom sheet
  - react-native-reanimated
  - react-native-gesture-handler
  - must install using `npx expo install package-name` to get compatible version for the SDK
- wrapped App.tsx with:
  - GestureHandlerRootView
  - BottomSheetProvider
