+++ src/navigation/FundDetailsStackNavigator.tsx
@@ -1,7 +1,7 @@
 import { createStackNavigator } from "@react-navigation/stack";
 
 import { AllocateFundScreen } from "../screens/funds/AllocateFundScreen";
-import { CashInScreen } from "../screens/funds/CashInScreen";
+import { CashInBottomSheet } from "../screens/funds/components/CashInBottomSheet";
 import { CashOutScreen } from "../screens/funds/CashOutScreen";
 import { FundDetailsScreen } from "../screens/funds/FundDetailsScreen";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
@@ -21,7 +21,7 @@ const Stack = createStackNavigator<FundDetailsStackParamList>();
 export const FundDetailsStackNavigator = () => (
   <Stack.Navigator screenOptions={{ headerShown: false }}>
     <Stack.Screen name="FundDetails" component={FundDetailsScreen} />
-    <Stack.Screen name="CashIn" component={CashInScreen} />
+    <Stack.Screen name="CashIn" component={CashInBottomSheet} />
     <Stack.Screen name="CashOut" component={CashOutScreen} />
     <Stack.Screen name="AllocateFund" component={AllocateFundScreen} />
   </Stack.Navigator>
