+++ src/navigation/FundStackNavigator.tsx
@@ -1,7 +1,7 @@
 import { createStackNavigator } from "@react-navigation/stack";
 
 import { AllocateFundScreen } from "../screens/funds/AllocateFundScreen";
-import { CashInScreen } from "../screens/funds/CashInScreen";
+import { CashInBottomSheet } from "../screens/funds/components/CashInBottomSheet";
 import { CashOutScreen } from "../screens/funds/CashOutScreen";
 import { FundsScreen } from "../screens/funds/FundsScreen";
 import {
@@ -14,7 +14,7 @@ import { NavigatorScreenParams } from "@react-navigation/native";
 export type FundsStackParamList = {
   Funds: undefined;
   FundDetailsStackNavigator: NavigatorScreenParams<FundDetailsStackParamList>;
-  CashIn: undefined;
+  // CashIn: undefined;
   CashOut: undefined;
   AllocateFund: undefined;
 };
@@ -26,11 +26,11 @@ const Stack = createStackNavigator<FundsStackParamList>();
 export const FundsStackNavigator = () => (
   <Stack.Navigator screenOptions={{ headerShown: false }}>
     <Stack.Screen name="Funds" component={FundsScreen} />
-    <Stack.Screen
+    {/* <Stack.Screen
       name="CashIn"
       component={CashInScreen}
       options={{ presentation: "modal" }}
-    />
+    /> */}
     <Stack.Screen
       name="CashOut"
       component={CashOutScreen}
