+++ src/screens/funds/FundsScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from "react";
+import React, { useEffect, useMemo, useRef } from "react";
 import { ScrollView } from "react-native";
 import { StyleSheet, TouchableOpacity, View } from "react-native";
 import { MaterialIcons, MaterialCommunityIcons } from "@expo/vector-icons";
@@ -16,19 +16,21 @@ import { useAppDispatch, useAppSelector } from "../../store/hooks";
 import { fetchFunds } from "../../store/funds/action";
 import { Funds } from "./components/Funds";
 import { LoadingScreen } from "../../components/Screens/LoadingScreen";
+import { BottomSheetModalMethods } from "@gorhom/bottom-sheet/lib/typescript/types";
+import { CashInBottomSheet } from "./components/CashInBottomSheet";
 
 export const FundsScreen = ({ navigation }: FundsStackProps) => {
   const dispatch = useAppDispatch();
   const { isFetching, funds } = useAppSelector((state) => state.fund);
-
-  const handleNavigateToCashInScreen = () => navigation.navigate("CashIn");
+  const sheetRef = useRef<BottomSheetModalMethods>(null);
+  // const handleNavigateToCashInScreen = () => navigation.navigate("CashIn");
   const handleNavigateToAllocateFundScreen = () =>
     navigation.navigate("AllocateFund");
   const handleNavigateToCashOutScreen = () => navigation.navigate("CashOut");
   const handleNavigateToFundDetailsScreen = () => navigation.navigate("Funds");
 
   useEffect(() => {
-    dispatch(fetchFunds());
+    // dispatch(fetchFunds());
   }, []);
 
   const totalFunds = useMemo(
@@ -49,7 +51,7 @@ export const FundsScreen = ({ navigation }: FundsStackProps) => {
               <Header>₱ {totalFunds.toLocaleString()}</Header>
             </View>
 
-            <TouchableOpacity onPress={handleNavigateToCashInScreen}>
+            <TouchableOpacity onPress={() => sheetRef.current?.present()}>
               <MaterialIcons
                 name="move-to-inbox"
                 size={32}
@@ -95,6 +97,7 @@ export const FundsScreen = ({ navigation }: FundsStackProps) => {
           </Padding>
         </ScrollView>
       </Padding>
+      <CashInBottomSheet ref={sheetRef} />
     </Screen>
   );
 };
